+++ assets/gulp/tasks/cssPost/postcss/build-css-post-components.js
@@ -9,7 +9,7 @@ const postcss = require("gulp-postcss");
 const uc = require("./util/unite-config");
 const asyncUtil = require("./util/async-util");
 const gutil = require("gulp-util");
-const cssnano = require("cssnano");
+const cssnano = require("gulp-cssnano");
 const errorUtil = require("./util/error-util");
 
 gulp.task("build-css-post-components", async () => {
@@ -31,7 +31,7 @@ gulp.task("build-css-post-components", async () => {
             errorCount++;
         })
         .on("error", errorUtil.handleErrorEvent)
-        .pipe(buildConfiguration.minify ? postcss([cssnano()]) : gutil.noop())
+        .pipe(buildConfiguration.minify ? cssnano() : gutil.noop())
         .pipe(buildConfiguration.sourcemaps ? sourcemaps.write({
             "includeContent": true,
             "sourceRoot": "./src"